Spring Validation 적용에 따라 코드 수정 중에 있습니다.
현재 소스는 에러 발생할 수 있습니다.

+++ NoticeService/src/main/java/kopo/poly/controller/NoticeController.java
@@ -4,7 +4,8 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import jakarta.servlet.http.HttpServletRequest;
+import jakarta.validation.Valid;
+import kopo.poly.controller.response.CommonResponse;
 import kopo.poly.dto.MsgDTO;
 import kopo.poly.dto.NoticeDTO;
 import kopo.poly.dto.TokenDTO;
@@ -13,16 +14,20 @@ import kopo.poly.service.ITokenAPIService;
 import kopo.poly.util.CmmUtil;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
-@CrossOrigin(origins = {"http://localhost:13000", "http://localhost:14000"}, allowedHeaders = {"POST, GET"}, allowCredentials = "true")
+@CrossOrigin(origins = {"http://localhost:13000", "http://localhost:14000"},
+        allowedHeaders = {"POST, GET"}, allowCredentials = "true")
 @Tag(name = "공지사항 서비스", description = "공지사항 구현을 위한 API")
 @Slf4j
-@RequestMapping(value = "/notice")
+@RequestMapping(value = "/notice/v1")
 @RequiredArgsConstructor
 @RestController
 public class NoticeController {
@@ -34,9 +39,10 @@ public class NoticeController {
 
     private final String HEADER_PREFIX = "Bearer "; // Bearer 토큰 사용을 위한 선언 값
 
-    @Operation(summary = "공지사항 리스트 API", description = "공지사항 리스트 정보 제공하는 API", responses = {@ApiResponse(responseCode = "200", description = "OK"), @ApiResponse(responseCode = "404", description = "Page Not Found!"),})
+    @Operation(summary = "공지사항 리스트 API", description = "공지사항 리스트 정보 제공하는 API",
+            responses = {@ApiResponse(responseCode = "200", description = "OK"), @ApiResponse(responseCode = "404", description = "Page Not Found!"),})
     @PostMapping(value = "noticeList")
-    public List<NoticeDTO> noticeList() {
+    public ResponseEntity<CommonResponse> noticeList() {
 
         // 로그 찍기(추후 찍은 로그를 통해 이 함수에 접근했는지 파악하기 용이하다.)
         log.info(this.getClass().getName() + ".noticeList Start!");
@@ -48,7 +54,8 @@ public class NoticeController {
         // 로그 찍기(추후 찍은 로그를 통해 이 함수 호출이 끝났는지 파악하기 용이하다.)
         log.info(this.getClass().getName() + ".noticeList End!");
 
-        return rList;
+        return ResponseEntity.ok(
+                CommonResponse.of(HttpStatus.OK, HttpStatus.OK.series().name(), rList));
     }
 
     @Operation(summary = "공지사항 상세보기 결과제공 API", description = "공지사항 상세보기 결과 및 조회수 증가 API",
@@ -57,28 +64,19 @@ public class NoticeController {
             responses = {@ApiResponse(responseCode = "200", description = "OK"),
                     @ApiResponse(responseCode = "404", description = "Page Not Found!"),})
     @PostMapping(value = "noticeInfo")
-    public NoticeDTO noticeInfo(HttpServletRequest request) throws Exception {
+    public ResponseEntity<CommonResponse> noticeInfo(@Valid @RequestBody NoticeDTO pDTO, BindingResult bindingResult) throws Exception {
 
         log.info(this.getClass().getName() + ".noticeInfo Start!");
 
-        String nSeq = CmmUtil.nvl(request.getParameter("nSeq")); // 공지글번호(PK)
-        String readCntYn = CmmUtil.nvl(request.getParameter("readCntYn")); // 조회수 증가여부
+        if (bindingResult.hasErrors()) { // Spring Validation 맞춰 잘 바인딩되었는지 체크
+            return CommonResponse.getErrors(bindingResult); // 유효성 검증 결과에 따른 에러 메시지 전달
 
-        boolean readCnt = readCntYn.equals("Y"); // 공지사항 증가여부를 boolean 값으로 변경
+        }
 
-        /*
-         * ####################################################################################
-         * 반드시, 값을 받았으면, 꼭 로그를 찍어서 값이 제대로 들어오는지 파악해야함 반드시 작성할 것
-         * ####################################################################################
-         */
-        log.info("nSeq : " + nSeq);
-        log.info("readCntYn : " + readCntYn);
-        log.info("readCnt : " + readCnt);
+        boolean readCnt = pDTO.readCntYn().equals("Y"); // 공지사항 증가여부를 boolean 값으로 변경
 
-        /*
-         * 값 전달은 반드시 DTO 객체를 이용해서 처리함 전달 받은 값을 DTO 객체에 넣는다.
-         */
-        NoticeDTO pDTO = NoticeDTO.builder().noticeSeq(Long.parseLong(nSeq)).build();
+        // 반드시, 값을 받았으면, 꼭 로그를 찍어서 값이 제대로 들어오는지 파악해야함 반드시 작성할 것
+        log.info("pDTO : " + pDTO);
 
         // Java 8부터 제공되는 Optional 활용하여 NPE(Null Pointer Exception) 처리
         NoticeDTO rDTO = Optional.ofNullable(noticeService.getNoticeInfo(pDTO, readCnt))
@@ -86,50 +84,47 @@ public class NoticeController {
 
         log.info(this.getClass().getName() + ".noticeInfo End!");
 
-        return rDTO;
+        return ResponseEntity.ok(
+                CommonResponse.of(HttpStatus.OK, HttpStatus.OK.series().name(), rDTO));
     }
 
     @Operation(summary = "공지사항 등록 API", description = "공지사항 등록 및 등록결과를 제공하는 API",
             responses = {@ApiResponse(responseCode = "200", description = "OK"),
                     @ApiResponse(responseCode = "404", description = "Page Not Found!"),})
     @PostMapping(value = "noticeInsert")
-    public MsgDTO noticeInsert(HttpServletRequest request,
-                               @CookieValue(value = "${jwt.token.access.name}") String token) {
+    public ResponseEntity<CommonResponse> noticeInsert(@Valid @RequestBody NoticeDTO pDTO, BindingResult bindingResult,
+                                                       @CookieValue(value = "${jwt.token.access.name}") String token) {
 
         log.info(this.getClass().getName() + ".noticeInsert Start!");
 
+        if (bindingResult.hasErrors()) { // Spring Validation 맞춰 잘 바인딩되었는지 체크
+            return CommonResponse.getErrors(bindingResult); // 유효성 검증 결과에 따른 에러 메시지 전달
+
+        }
+
         String msg = ""; // 메시지 내용
         int res = 0; // 성공 여부
-        MsgDTO dto = null; // 결과 메시지 구조
+        MsgDTO dto; // 결과 메시지 구조
 
         try {
-            TokenDTO tDTO = tokenAPIService.getTokenInfo(HEADER_PREFIX + token); // UserService로부터 Token 값 받아오기
+            // UserService로부터 Token 값 받아오기
+            TokenDTO tDTO = tokenAPIService.getTokenInfo(HEADER_PREFIX + token);
             log.info("TokenDTO : " + tDTO); // Token 값 출력하기
 
             //JWT Access 토큰으로부터 회원아이디 가져오기
             String userId = CmmUtil.nvl(tDTO.userId());
-            String title = CmmUtil.nvl(request.getParameter("title")); // 제목
-            String noticeYn = CmmUtil.nvl(request.getParameter("noticeYn")); // 공지글 여부
-            String contents = CmmUtil.nvl(request.getParameter("contents")); // 내용
 
-            /*
-             * ####################################################################################
-             * 반드시, 값을 받았으면, 꼭 로그를 찍어서 값이 제대로 들어오는지 파악해야함 반드시 작성할 것
-             * ####################################################################################
-             */
+            // 반드시, 값을 받았으면, 꼭 로그를 찍어서 값이 제대로 들어오는지 파악해야함 반드시 작성할 것
             log.info("userId : " + userId);
-            log.info("title : " + title);
-            log.info("noticeYn : " + noticeYn);
-            log.info("contents : " + contents);
+            log.info("pDTO : " + pDTO);
 
             // 데이터 저장하기 위해 DTO에 저장하기
-            NoticeDTO pDTO = NoticeDTO.builder().userId(userId).title(title)
-                    .noticeYn(noticeYn).contents(contents).build();
+            NoticeDTO nDTO = NoticeDTO.addUserId(pDTO, userId);
 
-            /*
-             * 게시글 등록하기위한 비즈니스 로직을 호출
-             */
-            noticeService.insertNoticeInfo(pDTO);
+            log.info("nDTO : " + nDTO);
+
+            // 게시글 등록하기위한 비즈니스 로직을 호출
+            noticeService.insertNoticeInfo(nDTO);
 
             // 저장이 완료되면 사용자에게 보여줄 메시지
             msg = "등록되었습니다.";
@@ -139,8 +134,8 @@ public class NoticeController {
 
             // 저장이 실패되면 사용자에게 보여줄 메시지
             msg = "실패하였습니다. : " + e.getMessage();
+            res = 0;
             log.info(e.toString());
-            e.printStackTrace();
 
         } finally {
             // 결과 메시지 전달하기
@@ -149,59 +144,55 @@ public class NoticeController {
             log.info(this.getClass().getName() + ".noticeInsert End!");
         }
 
-        return dto;
+        return ResponseEntity.ok(
+                CommonResponse.of(HttpStatus.OK, HttpStatus.OK.series().name(), dto));
     }
 
     @Operation(summary = "공지사항 수정 API", description = "공지사항 수정 및 수정결과를 제공하는 API",
             responses = {@ApiResponse(responseCode = "200", description = "OK"),
                     @ApiResponse(responseCode = "404", description = "Page Not Found!"),})
     @PostMapping(value = "noticeUpdate")
-    public MsgDTO noticeUpdate(HttpServletRequest request,
-                               @CookieValue(value = "${jwt.token.access.name}") String token) {
+    public ResponseEntity<CommonResponse> noticeUpdate(@Valid @RequestBody NoticeDTO pDTO, BindingResult bindingResult,
+                                                       @CookieValue(value = "${jwt.token.access.name}") String token) {
 
         log.info(this.getClass().getName() + ".noticeUpdate Start!");
 
+        if (bindingResult.hasErrors()) { // Spring Validation 맞춰 잘 바인딩되었는지 체크
+            return CommonResponse.getErrors(bindingResult); // 유효성 검증 결과에 따른 에러 메시지 전달
+
+        }
+
         String msg = ""; // 메시지 내용
         int res = 0; // 성공 여부
-        MsgDTO dto = null; // 결과 메시지 구조
+        MsgDTO dto; // 결과 메시지 구조
 
         try {
-            TokenDTO tDTO = tokenAPIService.getTokenInfo(HEADER_PREFIX + token); // UserService로부터 Token 값 받아오기
+            // UserService로부터 Token 값 받아오기
+            TokenDTO tDTO = tokenAPIService.getTokenInfo(HEADER_PREFIX + token);
             log.info("TokenDTO : " + tDTO); // Token 값 출력하기
 
+            //JWT Access 토큰으로부터 회원아이디 가져오기
             String userId = CmmUtil.nvl(tDTO.userId());
-            String nSeq = CmmUtil.nvl(request.getParameter("nSeq")); // 글번호(PK)
-            String title = CmmUtil.nvl(request.getParameter("title")); // 제목
-            String noticeYn = CmmUtil.nvl(request.getParameter("noticeYn")); // 공지글 여부
-            String contents = CmmUtil.nvl(request.getParameter("contents")); // 내용
 
-            /*
-             * ####################################################################################
-             * 반드시, 값을 받았으면, 꼭 로그를 찍어서 값이 제대로 들어오는지 파악해야함 반드시 작성할 것
-             * ####################################################################################
-             */
+            // 반드시, 값을 받았으면, 꼭 로그를 찍어서 값이 제대로 들어오는지 파악해야함 반드시 작성할 것
             log.info("userId : " + userId);
-            log.info("nSeq : " + nSeq);
-            log.info("title : " + title);
-            log.info("noticeYn : " + noticeYn);
-            log.info("contents : " + contents);
+            log.info("pDTO : " + pDTO);
 
-            /*
-             * 값 전달은 반드시 DTO 객체를 이용해서 처리함 전달 받은 값을 DTO 객체에 넣는다.
-             */
-            NoticeDTO pDTO = NoticeDTO.builder().userId(userId).noticeSeq(Long.parseLong(nSeq))
-                    .title(title).noticeYn(noticeYn).contents(contents).build();
+            // 데이터 저장하기 위해 DTO에 저장하기
+            NoticeDTO nDTO = NoticeDTO.addUserId(pDTO, userId);
+
+            log.info("nDTO : " + nDTO);
 
             // 게시글 수정하기 DB
-            noticeService.updateNoticeInfo(pDTO);
+            noticeService.updateNoticeInfo(nDTO);
 
             msg = "수정되었습니다.";
-            res = 0;
+            res = 1;
 
         } catch (Exception e) {
             msg = "실패하였습니다. : " + e.getMessage();
+            res = 0;
             log.info(e.toString());
-            e.printStackTrace();
 
         } finally {
 
@@ -212,35 +203,30 @@ public class NoticeController {
 
         }
 
-        return dto;
+        return ResponseEntity.ok(
+                CommonResponse.of(HttpStatus.OK, HttpStatus.OK.series().name(), dto));
     }
 
     @Operation(summary = "공지사항 삭제 API", description = "공지사항 삭제 및 삭제결과를 제공하는 API",
             responses = {@ApiResponse(responseCode = "200", description = "OK"),
                     @ApiResponse(responseCode = "404", description = "Page Not Found!"),})
     @PostMapping(value = "noticeDelete")
-    public MsgDTO noticeDelete(HttpServletRequest request) {
+    public ResponseEntity<CommonResponse> noticeDelete(@Valid @RequestBody NoticeDTO pDTO, BindingResult bindingResult) {
 
         log.info(this.getClass().getName() + ".noticeDelete Start!");
+
+        if (bindingResult.hasErrors()) { // Spring Validation 맞춰 잘 바인딩되었는지 체크
+            return CommonResponse.getErrors(bindingResult); // 유효성 검증 결과에 따른 에러 메시지 전달
+
+        }
 
         String msg = ""; // 메시지 내용
         int res = 0; // 성공 여부
         MsgDTO dto = null; // 결과 메시지 구조
 
         try {
-            String nSeq = CmmUtil.nvl(request.getParameter("nSeq")); // 글번호(PK)
-
-            /*
-             * ####################################################################################
-             * 반드시, 값을 받았으면, 꼭 로그를 찍어서 값이 제대로 들어오는지 파악해야함 반드시 작성할 것
-             * ####################################################################################
-             */
-            log.info("nSeq : " + nSeq);
-
-            /*
-             * 값 전달은 반드시 DTO 객체를 이용해서 처리함 전달 받은 값을 DTO 객체에 넣는다.
-             */
-            NoticeDTO pDTO = NoticeDTO.builder().noticeSeq(Long.parseLong(nSeq)).build();
+            // 반드시, 값을 받았으면, 꼭 로그를 찍어서 값이 제대로 들어오는지 파악해야함 반드시 작성할 것
+            log.info("pDTO : " + pDTO);
 
             // 게시글 삭제하기 DB
             noticeService.deleteNoticeInfo(pDTO);
@@ -250,8 +236,8 @@ public class NoticeController {
 
         } catch (Exception e) {
             msg = "실패하였습니다. : " + e.getMessage();
+            res = 0;
             log.info(e.toString());
-            e.printStackTrace();
 
         } finally {
             // 결과 메시지 전달하기
@@ -261,7 +247,8 @@ public class NoticeController {
 
         }
 
-        return dto;
+        return ResponseEntity.ok(
+                CommonResponse.of(HttpStatus.OK, HttpStatus.OK.series().name(), dto));
     }
 
 }
